requirements.txt
pyyaml
sendgrid
requests
html5lib
bs4
pyopenssl